fix():Serrialization model
