+++ crates/cloud-wallet-openid4vc/src/oid4vp/authorization/mod.rs
@@ -1,3 +1,5 @@
+mod error;
 mod response;
 
+pub use error::*;
 pub use response::*;
